Rust 2024: update Ch. 20 for new `unsafe` rules
- `extern` now requires `unsafe extern`, but in turn allows functions
  within it to be marked as `safe`. Add examples to both effects within
  the text, and update listing numbers accordingly.
- `#[no_mangle]` now requires `#[unsafe(no_mangle)]`

These are all backwards-compatible, so we can opt the chapter into using
the new rules without issue.
