+++ listings/ch20-advanced-features/listing-20-08/src/main.rs
@@ -1,4 +1,4 @@
-extern "C" {
+unsafe extern "C" {
     fn abs(input: i32) -> i32;
 }
