+++ listings/ch20-advanced-features/listing-20-09/src/main.rs
@@ -1,5 +1,7 @@
-static HELLO_WORLD: &str = "Hello, world!";
+unsafe extern "C" {
+    safe fn abs(input: i32) -> i32;
+}
 
 fn main() {
-    println!("name is: {HELLO_WORLD}");
+    println!("Absolute value of -3 according to C: {}", abs(-3));
 }
