+++ listings/ch20-advanced-features/listing-20-11/src/main.rs
@@ -1,9 +1,16 @@
-unsafe trait Foo {
-    // methods go here
+static mut COUNTER: u32 = 0;
+
+/// SAFETY: Calling this from more than a single thread at a time is undefined
+/// behavior, so you *must* guarantee you only call it from a single thread at
+/// a time.
+unsafe fn add_to_count(inc: u32) {
+    COUNTER += inc;
 }
 
-unsafe impl Foo for i32 {
-    // method implementations go here
+fn main() {
+    unsafe {
+        // SAFETY: This is only called from a single thread in `main`.
+        add_to_count(3);
+        println!("COUNTER: {}", COUNTER);
+    }
 }
-
-fn main() {}
