--- listings/ch20-advanced-features/listing-20-12/src/lib.rs
@@ -1,5 +1,0 @@
-pub trait Iterator {
-    type Item;
-
-    fn next(&mut self) -> Option<Self::Item>;
-}
