+++ listings/ch20-advanced-features/listing-20-12/src/main.rs
@@ -1,0 +1,9 @@
+unsafe trait Foo {
+    // methods go here
+}
+
+unsafe impl Foo for i32 {
+    // method implementations go here
+}
+
+fn main() {}
