+++ listings/ch20-advanced-features/listing-20-13/src/lib.rs
@@ -1,3 +1,5 @@
-pub trait Iterator<T> {
-    fn next(&mut self) -> Option<T>;
+pub trait Iterator {
+    type Item;
+
+    fn next(&mut self) -> Option<Self::Item>;
 }
