+++ listings/ch20-advanced-features/listing-20-14/src/lib.rs
@@ -1,0 +1,3 @@
+pub trait Iterator<T> {
+    fn next(&mut self) -> Option<T>;
+}
