--- listings/ch20-advanced-features/listing-20-18/output.txt
@@ -1,7 +1,0 @@
-$ cargo run
-   Compiling traits-example v0.1.0 (file:///projects/traits-example)
-    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.46s
-     Running `target/debug/traits-example`
-This is your captain speaking.
-Up!
-*waving arms furiously*
