+++ listings/ch20-advanced-features/listing-20-18/src/main.rs
@@ -29,8 +29,6 @@ impl Human {
 // ANCHOR: here
 fn main() {
     let person = Human;
-    Pilot::fly(&person);
-    Wizard::fly(&person);
     person.fly();
 }
 // ANCHOR_END: here
