+++ listings/ch20-advanced-features/listing-20-19/output.txt
@@ -1,5 +1,7 @@
 $ cargo run
    Compiling traits-example v0.1.0 (file:///projects/traits-example)
-    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.54s
+    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.46s
      Running `target/debug/traits-example`
-A baby dog is called a Spot
+This is your captain speaking.
+Up!
+*waving arms furiously*
