+++ listings/ch20-advanced-features/listing-20-19/src/main.rs
@@ -1,21 +1,36 @@
-trait Animal {
-    fn baby_name() -> String;
+trait Pilot {
+    fn fly(&self);
 }
 
-struct Dog;
+trait Wizard {
+    fn fly(&self);
+}
 
-impl Dog {
-    fn baby_name() -> String {
-        String::from("Spot")
+struct Human;
+
+impl Pilot for Human {
+    fn fly(&self) {
+        println!("This is your captain speaking.");
     }
 }
 
-impl Animal for Dog {
-    fn baby_name() -> String {
-        String::from("puppy")
+impl Wizard for Human {
+    fn fly(&self) {
+        println!("Up!");
     }
 }
 
+impl Human {
+    fn fly(&self) {
+        println!("*waving arms furiously*");
+    }
+}
+
+// ANCHOR: here
 fn main() {
-    println!("A baby dog is called a {}", Dog::baby_name());
+    let person = Human;
+    Pilot::fly(&person);
+    Wizard::fly(&person);
+    person.fly();
 }
+// ANCHOR_END: here
