+++ listings/ch20-advanced-features/listing-20-20/src/main.rs
@@ -16,8 +16,6 @@ impl Animal for Dog {
     }
 }
 
-// ANCHOR: here
 fn main() {
-    println!("A baby dog is called a {}", Animal::baby_name());
+    println!("A baby dog is called a {}", Dog::baby_name());
 }
-// ANCHOR_END: here
