+++ listings/ch20-advanced-features/listing-20-24/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "types-example"
+name = "traits-example"
 version = "0.1.0"
 edition = "2021"
