+++ listings/ch20-advanced-features/listing-20-24/src/main.rs
@@ -1,16 +1,14 @@
+use std::fmt;
+
+struct Wrapper(Vec<String>);
+
+impl fmt::Display for Wrapper {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "[{}]", self.0.join(", "))
+    }
+}
+
 fn main() {
-    // ANCHOR: here
-    let f: Box<dyn Fn() + Send + 'static> = Box::new(|| println!("hi"));
-
-    fn takes_long_type(f: Box<dyn Fn() + Send + 'static>) {
-        // --snip--
-    }
-
-    fn returns_long_type() -> Box<dyn Fn() + Send + 'static> {
-        // --snip--
-        // ANCHOR_END: here
-        Box::new(|| ())
-        // ANCHOR: here
-    }
-    // ANCHOR_END: here
+    let w = Wrapper(vec![String::from("hello"), String::from("world")]);
+    println!("w = {w}");
 }
