+++ listings/ch20-advanced-features/listing-20-25/src/main.rs
@@ -1,14 +1,12 @@
 fn main() {
     // ANCHOR: here
-    type Thunk = Box<dyn Fn() + Send + 'static>;
+    let f: Box<dyn Fn() + Send + 'static> = Box::new(|| println!("hi"));
 
-    let f: Thunk = Box::new(|| println!("hi"));
-
-    fn takes_long_type(f: Thunk) {
+    fn takes_long_type(f: Box<dyn Fn() + Send + 'static>) {
         // --snip--
     }
 
-    fn returns_long_type() -> Thunk {
+    fn returns_long_type() -> Box<dyn Fn() + Send + 'static> {
         // --snip--
         // ANCHOR_END: here
         Box::new(|| ())
