+++ listings/ch20-advanced-features/listing-20-26/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "unsafe-example"
+name = "types-example"
 version = "0.1.0"
 edition = "2021"
