+++ listings/ch20-advanced-features/listing-20-26/src/main.rs
@@ -1,0 +1,18 @@
+fn main() {
+    // ANCHOR: here
+    type Thunk = Box<dyn Fn() + Send + 'static>;
+
+    let f: Thunk = Box::new(|| println!("hi"));
+
+    fn takes_long_type(f: Thunk) {
+        // --snip--
+    }
+
+    fn returns_long_type() -> Thunk {
+        // --snip--
+        // ANCHOR_END: here
+        Box::new(|| ())
+        // ANCHOR: here
+    }
+    // ANCHOR_END: here
+}
