+++ listings/ch20-advanced-features/listing-20-28/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "macros-example"
+name = "functions-example"
 version = "0.1.0"
 edition = "2021"
