+++ listings/ch20-advanced-features/listing-20-29/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "functions-example"
+name = "macros-example"
 version = "0.1.0"
 edition = "2021"
