+++ listings/ch20-advanced-features/listing-20-31/src/main.rs
