+++ listings/ch20-advanced-features/listing-20-32/hello_macro/hello_macro_derive/src/lib.rs
