+++ listings/ch20-advanced-features/listing-20-32/hello_macro/src/lib.rs
