+++ listings/ch20-advanced-features/listing-20-34/hello_macro/src/lib.rs
